mark unused parameters to satisfy -Wextra
Parameters marked with the UNUSED_PARAMETERS macros cannot be
removed from functions because their prototypes cannot be changed.

Create the srv struct via designated initializers. Also remove NULL
assigment, because this way struct values are set to 0 anyway.

Also declare vars next to their first usage as usual.

Updates #443

+++ testserv.c
@@ -27,7 +27,7 @@ static byte fallocpat[3];
 static int
 wrapfalloc(int fd, int size)
 {
-    static int c = 0;
+    static size_t c = 0;
 
     printf("\nwrapfalloc: fd=%d size=%d\n", fd, size);
     if (c >= sizeof(fallocpat) || !fallocpat[c++]) {
